+++ changelog.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Changed
 
-- Define authoritative graph model (``)
+- Define authoritative graph model (`95f32aa`)
+- Detach demo from `svg-dom-graph` library (``)
 
 ## Added